Fix more issues where some binary expressions aren't sharded (#12921)
#### What this PR does

This PR is a follow up to https://github.com/grafana/mimir/pull/12900.
It fixes a number of other similar cases that affect other expressions
on one side of binary expressions. These were all caused by
`canShardAllVectorSelectors` incorrectly returning false.

I don't believe any of these would have caused incorrect query results,
but instead would have caused sharding not to be applied when it should
have been.

There is an opportunity to remove the "inspect only" mode introduced in
#12677 and instead make use of `willShardAllVectorSelectors` when
calculating the number of shardable legs, but I will do this in a follow
up PR.

I haven't added a changelog entry for the same reason I did not add one
in #12900.

#### Which issue(s) this PR fixes or relates to

#12900 and #12677

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [n/a] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.