Remote execution: buffer responses from queriers in query-frontends if needed (#12885)
#### What this PR does

This PR adds support for buffering responses from queriers in
query-frontends if the query contains sharded expressions and this
sharding is being performed inside MQE. This avoids a possible deadlock
scenario:

When concatenating results from multiple sharded legs, we need the
series metadata for all sharded legs before we can start streaming
series data back to the caller. If we don't eagerly load responses from
queriers, then in a query with (for example) 8 sharded legs, we might
end up deadlocked waiting for some or all of the last legs because these
requests can't start either due to the query-frontend's per-query max
parallelism limit or the number of available querier workers. If we
eagerly load responses, then exhausting the responses that have started
will eventually make space for the outstanding requests to complete.

⚠️ This PR is dependent on https://github.com/grafana/mimir/pull/12835.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12835

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.