Limit parallelism of sharded queries when sharding is running inside MQE (#12883)
#### What this PR does

This PR fixes the issue where the `max_query_parallelism` /
`-querier.max-query-parallelism` option is not respected if sharding is
running inside MQE.

⚠️ This PR is dependent on https://github.com/grafana/mimir/pull/12835
and https://github.com/grafana/mimir/pull/12885.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12835

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.