lookup planning: adjust plan's intersection cost (#12830)
### Problem

We've seen that ingesters can end up doing more work when this
optimization is enabled. This is due to fetching significantly more
series and checking them for sharding. This is not something our cost
model accounted for.

### What this PR does

this PR introduces the cost of retrieving a single series from the
index. The cost of doing this depends a lot on whether the block is an
in-memory or an on-disk block. In-memory blocks have much more efficient
sharding code. For now we're sticking with a single cost for all blocks,
but we can change that later.

### why 10?

I added a benchmark to help me come up with the new number. I'm
comparing the baseline of 2ns on my machine with the 2ns that it takes
to do a string comparison (see
https://github.com/grafana/mimir-prometheus/pull/990). This means our
cost should be around 10-40. 10 is 3 orders of magnitude higher than
what we had before (0.01) already, so I was cautious not to move this
too much


<details><summary>Cost of sharded vs non-sharded postings
iteration</summary>
<p>

false: not sharded
true: sharded

```
benchstat -col=/sharded -row=/size -filter='(/sharded:false OR /reuseCache:true)' 
        │    false     │                   true                   │
        │    sec/op    │    sec/op      vs base                   │
128       326.1n ± ∞ ¹   2937.0n ± ∞ ¹          ~ (p=0.100 n=3) ²
128K      301.5µ ± ∞ ¹   4943.5µ ± ∞ ¹          ~ (p=0.100 n=3) ²
1M        2.435m ± ∞ ¹   83.056m ± ∞ ¹          ~ (p=0.100 n=3) ²
geomean   62.10µ          1.064m        +1614.11%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05

        │    false     │                   true                   │
        │ sec/posting  │  sec/posting   vs base                   │
128       2.548n ± ∞ ¹   22.940n ± ∞ ¹          ~ (p=0.100 n=3) ²
128K      2.300n ± ∞ ¹   37.720n ± ∞ ¹          ~ (p=0.100 n=3) ²
1M        2.322n ± ∞ ¹   79.210n ± ∞ ¹          ~ (p=0.100 n=3) ²
geomean   2.387n          40.92n        +1614.16%
¹ need >= 6 samples for confidence interval at level 0.95
² need >= 4 samples to detect a difference at alpha level 0.05
```

</p>
</details> 



### Potential next steps
1. detect whether the block is in-memory or on-disk and adjust our
plans.
2. adjust the cost based on whether we have query sharding enabled or
not; hashing series is expensive

1\. will probably give higher returns


<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
3. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

#### Which issue(s) this PR fixes or relates to

related to https://github.com/grafana/mimir/issues/11920

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>