Add option to exclude workloads from `MimirRolloutStuck` alert (#12951)
#### What this PR does

This PR adds a configuration option to exclude particular StatefulSets
or Deployments from the `MimirRolloutStuck` alert.

This is useful to reduce false-positives for workloads that are expected
to be unavailable at times (eg. experimental workloads or those running
on spot nodes).

The default configuration preserves the existing behaviour of
considering all workloads.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.