Docs: update 'Grafana Mimir hash rings' page (#12924)
#### What this PR does

In this PR I've rewritten the 'Grafana Mimir hash rings' page, to try to
organise it more organically and explain how series sharding works both
in classic and ingest storage architecture.

#### Which issue(s) this PR fixes or relates to

Ref https://github.com/grafana/mimir-squad/issues/3161

#### Checklist

- [ ] Tests updated.
- [x] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>