Doc: update 'Configure Grafana Mimir hash rings' page (#12937)
#### What this PR does

In this PR I'm adding the mention to partitions ring to the 'Configure
Grafana Mimir hash rings' page. I also took the opportunity to update
all the links to use [the full URL format with version substitution
syntax](https://grafana.com/docs/writers-toolkit/write/links/#link-to-grafanacom-pages),
and to fix a couple of nits.

#### Which issue(s) this PR fixes or relates to

Ref: https://github.com/grafana/mimir-squad/issues/3161

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>