Query-frontend: more efficient expression-walking code (#12592)
Now that Prometheus exposes an iterator we can skip a couple of layers
of indirection and run just the code that we need.

This is more efficient in particular because `parser.Walk` would
allocate a slice for the expression path, which was never used.

The ability to stop with an error from `anyNode` was never used, and
`visitNode` was nearly identical, so elide into one.

Couple more related simplifications.