Fix issue where aggregation parameters aren't sharded even if the rest of the expression is shardable (#12958)
#### What this PR does

This PR fixes an issue where aggregation parameters aren't sharded even
if all other selectors in the expression are shardable.

For example, if the expression was `topk(scalar(sum(foo)), sum by (pod)
(bar))`, we would shard the `sum by (pod) (bar)`, but not the
`sum(foo)`.

Prior to https://github.com/grafana/mimir/pull/12785, this wouldn't be
an issue: the remaining `scalar(sum(foo))` would have been rewritten
into an embedded query by `subtreeFolder`, so it would still have been
evaluated without sharding. However, with
https://github.com/grafana/mimir/pull/12785, `subtreeFolder` has been
removed and so `scalar(sum(foo))` would have been left as-is, resulting
in the query failing with `shardedQuerier is expected embedded queries
but didn't find any`.

#### Which issue(s) this PR fixes or relates to

#12785

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.