Pin version of rollout-operator Jsonnet library (#12962)
#### What this PR does

This PR pins the version of the rollout-operator Jsonnet library to
https://github.com/grafana/rollout-operator/commit/4d563316583fdc8120c6ee7dfb393bb59a537bae,
which is v0.30.0 + https://github.com/grafana/rollout-operator/pull/309.

In the future, we should be able to use rollout-operator version tags
instead of commit hashes like this.

This prevents the Jsonnet tests from using rollout-operator `main` when
regenerating files, which can lead to unexpected diffs when new changes
are pushed to rollout-operator `main`.

#### Which issue(s) this PR fixes or relates to

(none)

#### Checklist

- [n/a] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.