Simplify sharding logic by removing "inspect only" mode (#12933)
#### What this PR does

This PR builds upon #12677, #12900 and #12921 to simplify some of the
logic introduced in #12677.

#12677 removed all instances of cloning the expression during sharding,
but this required us to introduce an "inspect only" mode to allow
`QuerySharder` to determine the correct number of shards without
mutating the expression.

This PR replaces this by building on the `willShardAllSelectors` method
introduced in the PRs above, as this method can calculate the number of
sharded selectors that will be returned in an expression, which removes
the need for "inspect only" mode.

#### Which issue(s) this PR fixes or relates to

#12677, #12900 and #12921

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.