Fix issue where step-invariant expressions aren't sharded correctly if preprocessing has been applied, and fix issue where unary expressions aren't sharded correctly in some situations (#12911)
#### What this PR does

This PR is a follow up to https://github.com/grafana/mimir/pull/12900
and fixes an issue where expressions containing step-invariant
expressions aren't sharded properly if preprocessing has been applied to
the query and the processed `parser.Expr` is passed to `shardSummer`.

This includes expressions like `sum(foo @ 123)`, as well as `sum(foo) >
(2*2)`.

This doesn't affect the middleware-based sharding implementation, as the
processed `parser.Expr` isn't passed to `shardSummer`, but it does
affect [the MQE
implementation](https://github.com/grafana/mimir/pull/12835).

This also fixes an issue where expressions containing unary negation
could be incorrectly sharded in some circumstances. This applies to both
implementations (and has been the case for a very long time).

I'd suggest reviewing each commit separately, d26b879 is a large diff
but this is just refactoring to make the following commit smaller.

⚠️ This PR builds upon #12933.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12835

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.