fsync write logs before offset is committed (#12816)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

This PR introduces the
`-ingest-storage.write-logs-fsync-before-kafka-commit-enabled` flag,
enabled by default, to make sure all WAL/WBL segment files are fsynced
before the corresponding offset is committed in Kafka.

Currently unclean ingester shutdowns can cause data loss, as write log
files aren't guaranteed to be fsynced before the corresponding offsets
are committed. This means that when an ingester restarts, it could see
corrupted WAL data (and therefore discard it), but not replay all the
required data from Kafka as it only resumes from the committed offset.

An additional
`-ingest-storage.write-logs-fsync-before-kafka-commit-concurrency` flag
has been added to enable us to reduce the amount of time to do the
fsyncs (and therefore the amount of time to commit an offset).

An optimisation is to only fsync tsdbs which have had updates since the
last offset was committed; we'll see how the current version runs in
production and decide whether that's worth adding it.

Depends on https://github.com/grafana/mimir-prometheus/pull/998

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features. (didn't update as
`-ingest-storage.*` is already mentioned as a whole)

---------

Co-authored-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>