Add min plan version to MQE planning nodes (#12961)
#### What this PR does

This PR introduces a MinimumRequiredPlanVersion version number to each
MQE query planning node.

Once a query plan has been built, the largest MinimumRequiredPlanVersion
found across all the nodes is used as the query plan's version.

Note that this version is encoded in the query plan when sent to remote
queriers, there by allowing queriers to reject plans which they do not
support.

#### Which issue(s) this PR fixes or relates to


#### Checklist

- [ x] Tests updated.
- [ x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.