Add support for specializing keep identifying OTel resource attributes setting (#12968)
#### What this PR does

Add support for specializing the _keep identifying OTel resource
attributes_ per-tenant parameter, in the same fashion as it works for
specializing the OTel resource attributes promotion per-tenant
parameter. This is to allow for other strategies than runtime overrides
of this particular configuration parameter.

#### Which issue(s) this PR fixes or relates to

#### Checklist

- [x] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Arve Knudsen <arve.knudsen@gmail.com>
Co-authored-by: Julius Hinze <julius.hinze@grafana.com>