jsonnet: Support setting distributor GOMEMLIMIT (#12963)
<!--  Thanks for sending a pull request!  Before submitting:

1. Read our CONTRIBUTING.md guide
2. Rebase your PR if it gets out of sync with main
-->

#### What this PR does

In Jsonnet library, support setting distributor `GOMEMLIMIT` equal to
its memory request, via `_config.distributor_gomemlimit_enabled`. If
enabled, the distributor horizontal auto-scaling memory trigger is also
removed, since it doesn't make sense in combination with `GOMEMLIMIT`.

We are already running with this configuration in Grafana Cloud.

#### Which issue(s) this PR fixes or relates to

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Arve Knudsen <arve.knudsen@gmail.com>