Jsonnet: expose functions to create memcached deployments (#12986)
#### What this PR does

I'm working to add multi-AZ support for the read path. As part of this
work I need to setup per-zone memcached. In this PR I'm doing a
refactoring to allow to create memcached deployments using functions.
Each function takes in input the name and node affinity (it's similar to
the args passed to other `newXXX()` functions we have for Mimir
components).

This refactoring is expected to be a no-op.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>