Docs: Update Config > Kafka to note that ingest storage is stable (#12977)
#### What this PR does
This PR updates the documentation for configuring the Kafka backend in
Grafana Mimir. The main change is updating the introductory explanation
to reflect that ingest storage is now the preferred and stable
architecture as of Mimir 3.0, and providing a clearer description of
Kafka's role in the architecture.

Documentation updates:

* Updated the introduction in `configure-kafka-backend.md` to clarify
that Kafka is used as the first layer of ingestion in the ingest storage
architecture, emphasizing improved scalability, decoupling, and
resilience. Also noted that as of Mimir 3.0, ingest storage is the
preferred and stable architecture.

#### Which issue(s) this PR fixes or relates to

Fixes https://github.com/grafana/mimir-squad/issues/3163

#### Checklist

- [ ] Tests updated.
- [X] Documentation added.
- [X] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.