Jsonnet: allow to customize node affinity matchers for memcached (#12987)
#### What this PR does

This PR is a follow up of https://github.com/grafana/mimir/pull/12986.
In this PR I'm proposing to add jsonnet fields to customise the
memcached's node affinity matchers. I've following the same design we
use for Mimir components, by exposing the fields at the root level.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Marco Pracucci <marco@pracucci.com>