MQE: Skip DeduplicateAndMerge if possible (#12528)
Introduce EliminateDeduplicateAndMerge optimization pass.

MQE wraps operations that could produce duplicate series (such as name-dropping functions, label manipulations, some binary operations) in DeduplicateAndMerge. However, when we can prove that the input series cannot contain duplicates, this step can be skipped. For example, rate(foo[5m]) cannot produce duplicates after rate drops the name, because all series with the name foo are guaranteed to be unique.
In contrast, for an expression like rate({__name__=~"(a|b)"}[5m]), the input series could produce duplicates after name is dropped.

EliminateDeduplicateAndMerge walks through the plan and collects DeduplicateAndMerge nodes. If it encounters a selector with an exact matcher, all collected DeduplicateAndMerge nodes are removed. If the selector uses a different type of matcher, only the DeduplicateAndMerge node closest to the selector is kept—once the name is dropped and the series are deduplicated, next name drops will not create new duplicates.
The exception is label_replace and label_join. Since they manipulate labels, they can generate duplicates or even reintroduce the __name__ label, making expressions prone to duplicates again. In these cases, we keep the DeduplicateAndMerge wrapping label_replace/label_join and the closest outermost node.

The main goal is to unblock label projections—the ability to request only the labels needed for a query. Since DeduplicateAndMerge requires all labels to perform deduplication and merging, label projections can only be used when deduplication is not required.

–––––––––––––––––––––––––––

Signed-off-by: Innokentii Konstantinov <innokenty.konstantinov@grafana.com>
Co-authored-by: Charles Korn <charleskorn@users.noreply.github.com>