Jsonnet: add functions to create ruler container and deployment (#12999)
#### What this PR does

Similar to https://github.com/grafana/mimir/pull/12986, in this PR I'm
adding functions to create the ruler container and deployment. I used
the same functions signature we already have for all other Mimir
components.

#### Which issue(s) this PR fixes or relates to

N/A

#### Checklist

- [ ] Tests updated.
- [ ] Documentation added.
- [ ] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Marco Pracucci <marco@pracucci.com>