chore: remove option to return HTTP 529 on rate limiting (#13012)
#### What this PR does

Remove an experimental flag to use HTTP 529 on rate limiting by the
distributor instead of the standard HTTP 429. The reason for having the
option (old Prometheus / Agents that didn't retry on 429) is no longer
an issue.

#### Which issue(s) this PR fixes or relates to

Related #5752

#### Checklist

- [X] Tests updated.
- [X] Documentation added.
- [X] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [X]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

Signed-off-by: Nick Pillitteri <nick.pillitteri@grafana.com>