MQE: Move out NameDrop functionality from DeduplicateAndMerge (#12848)
<!--  Thanks for sending a pull request!  Before submitting:
What this PR does
This PR introduces DropName operator, responsible for name label drop when DelayedNameRemoval is enabled.

Why?
It's a prerequisite to support DelayedNameRemoval in the EliminateDeduplicateAndMergeOptimizationPass
With the introduction of DelayedNameRemoval, DeduplicateAndMerge became responsible for dropping the name label. However, it conflicts with the idea of skipping DeduplicateAndMerge (PR). There are cases when we might skip DeduplicateAndMerge, but still need to drop a name label, so we can't do it. So, solution is to split them.

Example:
For the expression rate(foo[5m]) default plan is

- DeduplicateAndMerge
    - FunctionCall: rate(...)
        - MatrixSelector: {__name__="foo"}[5m0s]
selector foo guarantees that all series will be unique, so we can skip DeduplicateAndMerge. However, with DelayedNameRemoval enabled we can't eliminate it, because it's now dropping name.

After the change plan is:

- DeduplicateAndMerge
    - DropName
        - FunctionCall: rate(...)
           - MatrixSelector: {__name__="foo"}[5m0s]
And therefore DeduplicateAndMerge could be removed.

---------

Signed-off-by: Innokentii Konstantinov <innokenty.konstantinov@grafana.com>