Docs: Remove notes about mounting content to GEM docs (#12992)
This pull request removes a specific comment from multiple documentation
files in the Grafana Mimir docs. The comment previously reminded
maintainers that the topic is also used in GEM documentation and that
updates should be applicable to both. Since we aren't releasing GEM 3.0,
this comment is no longer applicable.

#### Which issue(s) this PR fixes or relates to

Fixes https://github.com/grafana/mimir-squad/issues/3301

#### Checklist

- [ ] Tests updated.
- [X] Documentation added.
- [X] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.