parquet-converter: allow multiple concurrent tasks (#12976)
#### What this PR does
Adds/implements two new config flags:
1. `parquet-converter.task-concurrency` (default 2). Determines how many
tasks from the queue are permitted to be executed concurrently
2. `parquet-converter.conversion-concurrency` (default 1). Determines
how many of the concurrent tasks are allowed to execute the
CPU-intensive block conversion code.

With these default values the converter will fall into a pattern of
downloading in one task while converting on the other.

---------

Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>