Docs: Go routines -> goroutines. (#13013)
#### What this PR does

This fixes a number of documentation references to "Go routines" to the
correct "goroutines."

(I am only opening this pedantic PR because "Go routines" is being
perpetuated for the sake of consistency. 😄 )

#### Checklist

- (N/A) Tests updated.
- [x] Documentation added.
- (N/A) `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- (N/A)
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.