Fix two more sharding bugs (#13027)
#### What this PR does

This PR fixes two issues in sharding.

##### Queries containing a function call where some arguments are
shardable and some are not fail

A query like `clamp_max(sum(foo), scalar(bar))` would fail with
`shardedQuerier is expecting embedded queries but didn't find any`.

This is another case similar to
https://github.com/grafana/mimir/pull/12958. Prior to
https://github.com/grafana/mimir/pull/12785, this wouldn't be an issue:
the remaining `scalar(bar)` would have been rewritten into an embedded
query by `subtreeFolder`, so it would still have been evaluated without
sharding. However, with https://github.com/grafana/mimir/pull/12785,
`subtreeFolder `has been removed and so `scalar(bar)` would have been
left as-is, resulting in the query failing with `shardedQuerier is
expected embedded queries but didn't find any`.

Unlike #12958, where the fix was to apply sharding to the aggregated
parameter, here the fix is to not shard the expression at all. A future
improvement would be to shard the shardable part of the expression and
retrieve the scalar arguments separately, but that's out of scope for
this fix.

##### Unshardable queries fail if the
`query_sharding_max_sharded_queries` limit is 0

Even with the fix above, queries like `scalar(bar)` would also fail with
`shardedQuerier is expecting embedded queries but didn't find any` if
the `query_sharding_max_sharded_queries` limit is 0.

This is because we weren't checking if the query was shardable if the
limit was 0, and so assumed the query should be sharded into the default
number of shards (`query_sharding_total_shards`).

Again, the removal of `subtreeFolder` in
https://github.com/grafana/mimir/pull/12785 uncovered the incorrect
logic here: previously, the unshardable expression would have been
rewritten into an embedded query by `subtreeFolder`, but with that
removed, the query was not being rewritten.

#### Which issue(s) this PR fixes or relates to

https://github.com/grafana/mimir/pull/12785

#### Checklist

- [x] Tests updated.
- [n/a] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [n/a]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

(cherry picked from commit 57a8e6ba98588109dbda3d61f50c9a8a7b8cee98)

# Conflicts:
#	CHANGELOG.md
#	pkg/frontend/querymiddleware/astmapper/sharding_test.go
#	pkg/frontend/querymiddleware/querysharding.go