feat(query-tee): Support query-tee comparisons for multiple secondary backends (#13022)
Although the query-tee had support for multiple secondary backends, it
did not support comparing results of the preferred backend against all
secondaries. This PR makes it possible and slightly updates the service
instrumentation to report results.

#### Checklist

- [x] Tests updated.
- [x] Documentation added.
- [x] `CHANGELOG.md` updated - the order of entries should be
`[CHANGE]`, `[FEATURE]`, `[ENHANCEMENT]`, `[BUGFIX]`. If changelog entry
is not needed, please add the `changelog-not-needed` label to the PR.
- [ ]
[`about-versioning.md`](https://github.com/grafana/mimir/blob/main/docs/sources/mimir/configure/about-versioning.md)
updated with experimental features.

---------

Signed-off-by: Jesus Vazquez <jesus.vazquez@grafana.com>
Co-authored-by: Taylor C <41653732+tacole02@users.noreply.github.com>