helm: remove GEM and enterprise-releated resources (#13005)
This PR includes changes reviewed separately in a few PRs:


* https://github.com/grafana/mimir/pull/13005
* https://github.com/grafana/mimir/pull/13006
* https://github.com/grafana/mimir/pull/13007
* https://github.com/grafana/mimir/pull/13008
* https://github.com/grafana/mimir/pull/13009
* https://github.com/grafana/mimir/pull/13010





### https://github.com/grafana/mimir/pull/13005

this is a preparation for removing GEM from the helm chart. This removes
some rendered helm tests. This makes it easier to review the diffs of
later PRs


---

### https://github.com/grafana/mimir/pull/13006

based on #13005 this deletes any components which require a GEM license
to run. It doesn't take care of removing the GEM gateway from
operations/helm/charts/mimir-distributed/templates/gateway/gateway-dep.yaml,
but that will come in a follow-up PR. A follow-up PR will also remove
the keys from values.yaml


---

### https://github.com/grafana/mimir/pull/13007

based on #13006, this deletes some scripts related to testing graphite
and generating GEM tokens

---

### https://github.com/grafana/mimir/pull/13008

based on https://github.com/grafana/mimir/pull/13007 this removes the
license volume on all components


---
### https://github.com/grafana/mimir/pull/13009

based on #13008 this removes any usage of enterprise-related values from
any template in the chart

the idea is to show that in reality very little changes in our rendered
manifests for non-enterprise charts (see offline tests). The two offline
tests that change have enterprise enabled



---

### https://github.com/grafana/mimir/pull/13010

based on #13009 this removes any references to GEM and GEM components
from the values files in the chart. Probably the biggest change here is
that we're removing the buckets from the minio setup config

---------

Signed-off-by: Dimitar Dimitrov <dimitar.dimitrov@grafana.com>